Refactoring: Establish `outline.core` module
