+++ src/cratedb_about/cli.py
@@ -6,7 +6,7 @@ import click
 from pueblo.util.cli import boot_click
 
 from cratedb_about.build.llmstxt import LllmsTxtBuilder
-from cratedb_about.outline.model import CrateDbKnowledgeOutline
+from cratedb_about.outline import CrateDbKnowledgeOutline
 from cratedb_about.query.core import CrateDbKnowledgeConversation
 from cratedb_about.query.model import Example
