+++ src/cratedb_about/outline/__init__.py
@@ -1,5 +1,7 @@
-from .model import CrateDbKnowledgeOutline
+from .core import CrateDbKnowledgeOutline
+from .model import OutlineDocument
 
 __all__ = [
     "CrateDbKnowledgeOutline",
+    "OutlineDocument",
 ]
